requirements.txt
neo4j
requests
networkx
matplotlib
python-dotenv